requirements.txt
streamlit
newspaper3k
transformers
gspread
oauth2client
validators
lxml_html_clean
torch
